The website
Just made a dummy website in about 2 hours. I'll update it to look better when I have time, for now just wanted to get something up there.
It follows this:  THE WEBSITE – Must be Public
-	The website must have a home page that displays all of the following:
o	Your project ID, color (if any) and title
o	Your bio pictures and names – bio pics do not need to be huge
o	Course number, semester and year
o	Navigation links to other documents (as pdf not word !)
o	Links must be working (DO NOT LINK TO GOOGLE DOCS – create a report as a pdf for your website)
o	Navigation link to your Final Presentation Video – that is narrated, explains your project development approach and demos your final project
o	Everything should be visible on the page – do not make me SCROLL                                                                                                                                                                                                                                                                                              
hopefully the minimal scrolling wont be a huge deal. If it is I'll find ways tonight to make it scaleable and fit on one page.